Add files via upload
Task description: Write a program to check whether or not a given number is strong. A number is strong if the sum of the Factorial of each digit is equal to the number. For example 145 is a strong number, because 1! + 4! + 5! = 145. Print "yes" if the number is strong and "no" if the number is not strong.
Examples:
Input	Output
2	        yes
3451	no
40585	yes